Use default rosbag options
